+++ pkgs/firehose/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.20-wip
 - Cache file contents and parsed data in `ChangeLog` class.
+- Add code coverage to checks.
 
 ## 0.3.19
 - Clean-up and optimizations.